backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0
drf-nested-routers==0.94.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
openai==1.57.4
stripe==11.3.0
requests==2.32.3
celery==5.4.0
redis==5.2.0
cryptography==44.0.0
Pillow==11.0.0
gunicorn==23.0.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
openpyxl==3.1.5
user-agents==2.2.0
scipy==1.14.1
django-storages==1.14.4
boto3==1.35.76
twilio==9.8.8
pyjwt==2.10.1
django-redis==5.4.0

# New dependencies for advanced features
pytz==2024.2
pyotp==2.9.0
qrcode==8.0
Jinja2==3.1.4
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
pywebpush==1.14.1
ipaddress==1.0.23
